# shared/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/schedule/{event_id}/current": {
      "patch": {
        "tags": [
          "Schedule"
        ],
        "summary": "Set specific event as current",
        "description": "Updates the schedule state to mark a specific event as active. Validates timing and event status.",
        "operationId": "set_event_as_current_schedule__event_id__current_patch",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Schedule event ID.",
              "title": "Event Id"
            },
            "description": "Schedule event ID."
          }
        ],
        "responses": {
          "200": {
            "description": "Event set as current successfully.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Request validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "400": {
            "description": "Event is skipped or invalid for this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Event ID not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited — editing too fast.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds to wait before retrying."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/schedule/current": {
      "delete": {
        "tags": [
          "Schedule"
        ],
        "summary": "Unset current schedule event",
        "description": "Clears the currently active event from the schedule. Subject to rate limiting.",
        "operationId": "uncheck_current_event_schedule_current_delete",
        "responses": {
          "200": {
            "description": "Current event cleared successfully.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Request validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited — editing too fast.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/schedule/{event_id}/move": {
      "patch": {
        "tags": [
          "Schedule"
        ],
        "summary": "Reorder schedule event",
        "description": "Moves an event to a new position in the sequence, specifically after the provided event ID.",
        "operationId": "move_schedule_event_schedule__event_id__move_patch",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Schedule event ID.",
              "title": "Event Id"
            },
            "description": "Schedule event ID."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveScheduleEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event moved successfully.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Request validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid move: target and destination are the same.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Event to move or target neighbor event was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/schedule/{event_id}/skip": {
      "patch": {
        "tags": [
          "Schedule"
        ],
        "summary": "Skip a schedule event",
        "description": "Marks a specific event as skipped. Note: the currently active event cannot be skipped.",
        "operationId": "skip_schedule_event_schedule__event_id__skip_patch",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Schedule event ID.",
              "title": "Event Id"
            },
            "description": "Schedule event ID."
          }
        ],
        "responses": {
          "200": {
            "description": "Event marked as skipped.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Request validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "400": {
            "description": "Cannot skip the current active event.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Event ID not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/schedule/{event_id}/unskip": {
      "patch": {
        "tags": [
          "Schedule"
        ],
        "summary": "Unskip a schedule event",
        "description": "Restores a previously skipped event back into the active schedule sequence.",
        "operationId": "unskip_schedule_event_schedule__event_id__unskip_patch",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Schedule event ID.",
              "title": "Event Id"
            },
            "description": "Schedule event ID."
          }
        ],
        "responses": {
          "200": {
            "description": "Event restored to active schedule.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Request validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "400": {
            "description": "Event state cannot be modified in its current context.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Event ID not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorMessage": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "ErrorMessage"
      },
...
      "MoveScheduleEventRequest": {
        "properties": {
          "place_after_event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Place After Event Id"
          }
        },
        "type": "object",
        "required": [
          "place_after_event_id"
        ],
        "title": "MoveScheduleEventRequest"
      },
...
      "ValidationErrorDetail": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Loc"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "type"
        ],
        "title": "ValidationErrorDetail"
      },
      "ValidationErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "VALIDATION_ERROR"
          },
          "details": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/ValidationErrorDetail"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "title": "ValidationErrorResponse"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "description": "Session cookie set after login.",
        "in": "cookie",
        "name": "session_id"
      }
    }
  }
}